Move User model file to ./app/Models
